sit_be/requirements.txt
arabic-reshaper==3.0.0
asn1crypto==1.5.1
bcrypt==4.0.1
blinker==1.6.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.3
clr-loader==0.2.6
colorama==0.4.6
cryptography==40.0.2
cssselect2==0.7.0
dnspython==2.3.0
Flask==2.3.2
Flask-Cors==3.0.10
Flask-Mail==0.9.1
Flask-PyMongo==2.3.0
gunicorn==21.2.0
html5lib==1.1
idna==3.4
IronPdf==2023.8.6
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.3
MarkupSafe==2.1.2
oscrypto==1.3.0
packaging==23.1
pdfkit==1.0.0
Pillow==10.0.0
pycparser==2.21
pyHanko==0.20.0
pyhanko-certvalidator==0.23.0
PyJWT==2.6.0
pymongo==4.3.3
pypdf==3.15.5
pypng==0.20220715.0
python-bidi==0.4.2
python-dotenv==1.0.0
pythonnet==3.0.2
pytz==2023.3
PyYAML==6.0.1
qrcode==7.4.2
reportlab==3.6.13
requests==2.31.0
six==1.16.0
svglib==1.5.1
tinycss2==1.2.1
typing_extensions==4.7.1
tzlocal==5.0.1
uritools==4.0.2
urllib3==2.0.4
webencodings==0.5.1
Werkzeug==2.3.3
xhtml2pdf==0.2.11
